feat(backend): move functino from weekly to monthly
