+++ hapi/src/cron-jobs/monthly.js
@@ -1,5 +1,5 @@
 const mailApi = require('../utils/mail')
-const { sponsorApi, lifebankApi, userApi, offerApi } = require('../api')
+const { sponsorApi, lifebankApi, userApi, offerApi, accountApi } = require('../api')
 
 const generateSponsorsTransactionReports = async () => {
   const today = new Date()
@@ -121,29 +121,71 @@ const generateLifebanksTransactionReports = async () => {
   })
 }
 
-const sendEmail = async () => {
-  try {
-    await mailApi.sendConfirmMessage(
-      'leisterac.1997@gmail.com',
-      'MONTHLY',
-      'TITLE',
-      'MESSAGE'
+const generateNewSponsorAndOfferReportToLifebanks = async () => {
+  const today = new Date()
+  const monthAgo = new Date()
+  monthAgo.setDate(monthAgo.getDate() - 7)
+
+  const lifebanks = await userApi.getMany({
+    role: { _eq: 'lifebank' }
+  })
+
+  const newSponsors = await userApi.getMany({
+    role: { _eq: 'sponsor' },
+    created_at: { _gte: monthAgo, _lte: today }
+  })
+
+  const newOffers = await offerApi.getMany({
+    created_at: { _gte: monthAgo, _lte: today }
+  })
+
+  lifebanks.forEach(async (lifebank) => {
+    const lifebankProfile = await accountApi.getProfile(lifebank.account)
+
+    const nerbySponsors = []
+    for (i = 0; i < newSponsors.length; i++) {
+      const tempSponsor = newSponsors[i]
+      const sponsorProfile = await accountApi.getProfile(tempSponsor.account)
+      sponsorProfile.location &&
+      accountApi.isCoordinateInsideBox(
+        JSON.parse(sponsorProfile.location),
+        JSON.parse(lifebankProfile.location)
+      )
+        ? nerbySponsors.push(tempSponsor)
+        : null
+    }
+
+    let stringSponsorHtmlContent = ''
+    nerbySponsors.forEach((sponsor) => {
+      stringSponsorHtmlContent = stringSponsorHtmlContent.concat(
+        '<br><a href="',
+        'https://lifebank.io/info/',
+        sponsor.account,
+        '">',
+        sponsor.name,
+        '</a>'
+      )
+    })
+
+    let stringOfferHtmlContent = ''
+    newOffers.forEach((offer) => {
+      stringOfferHtmlContent = stringOfferHtmlContent.concat(
+        '<br><a href="https://lifebank.io/">',
+        offer.offer_name,
+        '</a>'
+      )
+    })
+
+    mailApi.sendNewSponsorAndOfferReport(
+      lifebank.email,
+      lifebank.language,
+      lifebank.role,
+      stringSponsorHtmlContent,
+      stringOfferHtmlContent
     )
-
-    return {
-      success: true
-    }
-  } catch (error) {
-    console.log(error)
-    return {
-      success: false
-    }
-  }
+  })
 }
 
-// generateSponsorsTransactionReports()
-// generateLifebanksTransactionReports()
-module.exports = {
-  generateSponsorsTransactionReports,
-  generateLifebanksTransactionReports
-}
+generateSponsorsTransactionReports()
+generateLifebanksTransactionReports()
+generateNewSponsorAndOfferReportToLifebanks()
