+++ hapi/src/routes/index.js
@@ -23,7 +23,6 @@ const getReport = require('./get-report/get-report.route')
 const redeemOffer = require('./redeem-offer/redeem-offer.route')
 const checkEmailVerified = require('./check-email-verified/check-email-verified.route')
 const donate = require('./donate/donate.route')
-const test = require('./test/test.route')
 
 module.exports = [
   changePasswordRoute,
@@ -51,5 +50,4 @@ module.exports = [
   redeemOffer,
   checkEmailVerified,
   donate,
-  test
 ]
