--- hapi/src/routes/test/test.route.js
@@ -1,10 +1,0 @@
-const testHandler = require('./test.handler')
-
-module.exports = {
-  method: 'POST',
-  path: '/api/test',
-  handler: testHandler,
-  options: {
-    auth: false
-  }
-}
